Pathfinding Implemented
Basic pathfinding is implemented
the map now  has nodes throughout and the enemy class finds a path to
the player and only recalculates once it has walked through it.
There are some problems like longer paths than there should be and
collision for enemies between walls has been disabled temporarily. The
pathfinder still does not know the location of other entities so things
like doors and liftables are not obstacles yet.